@@ -1,29 +1,28 @@
 module github.com/grafana/plugin-validator
 
-go 1.24.6
-
-toolchain go1.25.4
+go 1.25.5
 
 require (
-	github.com/bmatcuk/doublestar/v4 v4.9.1
+	github.com/bmatcuk/doublestar/v4 v4.9.2
 	github.com/danwakefield/fnmatch v0.0.0-20160403171240-cbb64ac3d964
 	github.com/fatih/color v1.18.0
 	github.com/go-enry/go-license-detector/v4 v4.3.1
 	github.com/google/generative-ai-go v0.20.1
-	github.com/google/osv-scanner/v2 v2.2.4
-	github.com/hashicorp/go-version v1.7.0
+	github.com/google/osv-scanner/v2 v2.3.1
+	github.com/hashicorp/go-version v1.8.0
 	github.com/jarcoal/httpmock v1.4.1
 	github.com/magefile/mage v1.15.0
-	github.com/ossf/osv-schema/bindings/go v0.0.0-20251012234424-434020c6442f
+	github.com/ossf/osv-schema/bindings/go v0.0.0-20251230224438-88c48750ddae
 	github.com/r3labs/diff/v3 v3.0.2
 	github.com/smartystreets/goconvey v1.8.1
 	github.com/sourcegraph/go-diff-patch v0.0.0-20240223163233-798fd1e94a8e
 	github.com/stretchr/testify v1.11.1
 	github.com/tailscale/hujson v0.0.0-20250605163823-992244df8c5a
 	github.com/xeipuuv/gojsonschema v1.2.0
-	golang.org/x/crypto v0.45.0
-	golang.org/x/mod v0.29.0
-	google.golang.org/api v0.255.0
+	golang.org/x/crypto v0.46.0
+	golang.org/x/mod v0.31.0
+	google.golang.org/api v0.258.0
+	google.golang.org/protobuf v1.36.11
 	gopkg.in/yaml.v3 v3.0.1
 )
 
