@@ -7,7 +7,22 @@ import (
 	"strings"
 
 	"github.com/grafana/plugin-validator/pkg/analysis"
+	"github.com/grafana/plugin-validator/pkg/analysis/passes/archive"
+	"github.com/grafana/plugin-validator/pkg/analysis/passes/backendbinary"
+	"github.com/grafana/plugin-validator/pkg/analysis/passes/binarypermissions"
+	"github.com/grafana/plugin-validator/pkg/analysis/passes/coderules"
+	"github.com/grafana/plugin-validator/pkg/analysis/passes/gomanifest"
+	"github.com/grafana/plugin-validator/pkg/analysis/passes/jssourcemap"
+	"github.com/grafana/plugin-validator/pkg/analysis/passes/manifest"
+	"github.com/grafana/plugin-validator/pkg/analysis/passes/metadata"
+	"github.com/grafana/plugin-validator/pkg/analysis/passes/metadatavalid"
+	"github.com/grafana/plugin-validator/pkg/analysis/passes/modulejs"
+	"github.com/grafana/plugin-validator/pkg/analysis/passes/osvscanner"
+	"github.com/grafana/plugin-validator/pkg/analysis/passes/safelinks"
 	"github.com/grafana/plugin-validator/pkg/analysis/passes/sourcecode"
+	"github.com/grafana/plugin-validator/pkg/analysis/passes/trackingscripts"
+	"github.com/grafana/plugin-validator/pkg/analysis/passes/unsafesvg"
+	"github.com/grafana/plugin-validator/pkg/analysis/passes/virusscan"
 	"github.com/grafana/plugin-validator/pkg/llmvalidate"
 	"github.com/grafana/plugin-validator/pkg/logme"
 )
@@ -15,15 +30,45 @@ import (
 var geminiKey = os.Getenv("GEMINI_API_KEY")
 
 var (
-	llmIssueFound   = &analysis.Rule{Name: "llm-issue-found", Severity: analysis.SuspectedProblem}
-	llmWarningFound = &analysis.Rule{Name: "llm-warning-found", Severity: analysis.Warning}
+	llmIssueFound    = &analysis.Rule{Name: "llm-issue-found", Severity: analysis.SuspectedProblem}
+	llmWarningFound  = &analysis.Rule{Name: "llm-warning-found", Severity: analysis.Warning}
+	llmReviewSkipped = &analysis.Rule{
+		Name:     "llm-review-skipped",
+		Severity: analysis.SuspectedProblem,
+	}
 )
+
+// blockingAnalyzers contains validators that, if they report errors, should cause
+// the LLM review to be skipped to save costs. These are grouped into:
+// - Tier 1 (Structure): archive, metadata, metadatavalid, modulejs
+// - Tier 2 (Security): coderules, trackingscripts, virusscan, safelinks, unsafesvg, osvscanner
+// - Tier 3 (Integrity): gomanifest, binarypermissions, backendbinary, manifest
+var blockingAnalyzers = []*analysis.Analyzer{
+	// Tier 1: Fundamental structure issues
+	archive.Analyzer,
+	metadata.Analyzer,
+	metadatavalid.Analyzer,
+	modulejs.Analyzer,
+	// Tier 2: Security/Policy violations
+	coderules.Analyzer,
+	trackingscripts.Analyzer,
+	virusscan.Analyzer,
+	safelinks.Analyzer,
+	unsafesvg.Analyzer,
+	osvscanner.Analyzer,
+	// Tier 3: Build/Integrity issues
+	gomanifest.Analyzer,
+	jssourcemap.Analyzer,
+	binarypermissions.Analyzer,
+	backendbinary.Analyzer,
+	manifest.Analyzer,
+}
 
 var Analyzer = &analysis.Analyzer{
 	Name:     "llmreview",
-	Requires: []*analysis.Analyzer{sourcecode.Analyzer},
+	Requires: append([]*analysis.Analyzer{sourcecode.Analyzer}, blockingAnalyzers...),
 	Run:      run,
-	Rules:    []*analysis.Rule{llmIssueFound},
+	Rules:    []*analysis.Rule{llmIssueFound, llmReviewSkipped},
 	ReadmeInfo: analysis.ReadmeInfo{
 		Name:         "LLM Review",
 		Description:  "Runs the code through Gemini LLM to check for security issues or disallowed usage.",
@@ -70,10 +115,6 @@ var Questions = []llmvalidate.LLMQuestion{
 	},
 	{
 		Question:       "Only for go/golang code: Does this code create HTTP clients without using github.com/grafana/grafana-plugin-sdk-go/backend/httpclient? (Look for direct creation of http.Client{}, http.NewRequest, calls to third-party NewClient/NewHTTPClient functions that don't accept or use the SDK's httpclient, or any other HTTP client initialization that doesn't use github.com/grafana/grafana-plugin-sdk-go/backend/httpclient. The httpclient from github.com/grafana/grafana-plugin-sdk-go/backend/httpclient should be used directly or passed to the HTTP client being created. This includes cases where third-party libraries create HTTP clients internally - those libraries should accept the SDK's httpclient as a parameter). Provide the specific code snippet if found.",
-		ExpectedAnswer: false,
-	},
-	{
-		Question:       "Does this code contain console.log(), console.warn(), console.info(), or console.debug() statements that would output in production builds? (Exclude code wrapped in NODE_ENV checks or test files). Provide the specific code snippet if found.",
 		ExpectedAnswer: false,
 	},
 	{
@@ -171,6 +212,23 @@ func run(pass *analysis.Pass) (any, error) {
 		return nil, nil
 	}
 
+	// Check if any blocking analyzers reported errors - skip LLM review to save costs
+	// keep here before source code and key check for tests to work
+	for _, analyzer := range blockingAnalyzers {
+		if pass.AnalyzerHasErrors(analyzer) {
+			pass.ReportResult(
+				pass.AnalyzerName,
+				llmReviewSkipped,
+				fmt.Sprintf("LLM review skipped due to errors in %s", analyzer.Name),
+				fmt.Sprintf(
+					"Fix the errors reported by %s before LLM review can run.",
+					analyzer.Name,
+				),
+			)
+			return nil, nil
+		}
+	}
+
 	var err error
 	// only run if sourcecode.Analyzer succeeded
 	sourceCodeDir, ok := pass.ResultOf[sourcecode.Analyzer].(string)
@@ -184,7 +242,7 @@ func run(pass *analysis.Pass) (any, error) {
 
 	logme.Debugln("Starting to run Gemini Validations. This might take a while...")
 
-	llmClient, err := llmvalidate.New(context.Background(), geminiKey, "gemini-2.5-flash")
+	llmClient, err := llmvalidate.New(context.Background(), geminiKey, "gemini-3-flash-preview")
 
 	if err != nil {
 		logme.DebugFln("Error initializing llm client: %v", err)
@@ -241,12 +299,20 @@ func buildDetailString(answer llmvalidate.LLMAnswer) string {
 	detailParts = append(detailParts, answer.Answer)
 
 	if answer.CodeSnippet != "" {
-		detailParts = append(detailParts, fmt.Sprintf("**Code Snippet:**\n```\n%s\n```", answer.CodeSnippet))
+		detailParts = append(
+			detailParts,
+			fmt.Sprintf("**Code Snippet:**\n```\n%s\n```", answer.CodeSnippet),
+		)
 	}
 
 	if len(answer.Files) > 0 {
-		detailParts = append(detailParts, fmt.Sprintf("**Files:** %s", strings.Join(answer.Files, ", ")))
+		detailParts = append(
+			detailParts,
+			fmt.Sprintf("**Files:** %s", strings.Join(answer.Files, ", ")),
+		)
 	}
 
-	return strings.Join(detailParts, "\n\n")
+	detail := strings.Join(detailParts, "\n\n")
+
+	return detail
 }
