@@ -60,8 +60,7 @@ type MetadataDependencies struct {
 }
 
 type MetadataPluginDependency struct {
-	ID      string `json:"id"`
-	Name    string `json:"name"`
-	Type    string `json:"type"`
-	Version string `json:"version"`
+	ID   string `json:"id"`
+	Name string `json:"name"`
+	Type string `json:"type"`
 }
